pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["common*", "trader*", "api*", "scripts*"]

[project]
name = "market-ai"
version = "0.1.0"
description = "Automated swing trading bot for Interactive Brokers"
requires-python = ">=3.11"
dependencies = [
    "ib_insync>=0.9.86",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "sqlalchemy>=2.0",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "feedparser>=6.0",
    "numpy>=1.26",
    "pandas>=2.2",
    "psycopg[binary]>=3.1",
    "alembic>=1.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
